src/main/java/lloyd/hapag/mecka/franciszek/loginapp/user/service/UserService.java
package lloyd.hapag.mecka.franciszek.loginapp.user.service;

import lloyd.hapag.mecka.franciszek.loginapp.user.dto.UserDto;
import lloyd.hapag.mecka.franciszek.loginapp.user.entity.User;
import lloyd.hapag.mecka.franciszek.loginapp.user.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.modelmapper.ModelMapper;

import java.util.NoSuchElementException;
import java.util.Optional;

@Service
public class UserService {
    private final UserRepository repository;

    private final ModelMapper modelMapper;

    public UserService(UserRepository repository, ModelMapper modelMapper) {
        this.repository = repository;
        this.modelMapper = modelMapper;
    }

    public UserDto convertToDto(User user) {
        return modelMapper.map(user, UserDto.class);
    }

    public User convertToEntity(UserDto userDto) {
        return modelMapper.map(userDto, User.class);
    }

    public Optional<User> find(Long id) {
        return repository.findById(id);
    }

    public User create(User user) {
        return repository.save(user);
    }

    public User update(UserDto userDto, long id) {
        User user = repository.findById(id).orElseThrow(NoSuchElementException::new);

        user.setAge(userDto.getAge());
        user.setGender(userDto.getGender());


        return repository.save(user);
    }

    public void delete(Long id) {
        repository.findById(id).ifPresent(repository::delete);
    }

}
